+++ src/model.rs
@@ -14,9 +14,6 @@
 
 pub mod app_state;
 pub mod document;
-pub mod mod_record;
-pub mod mod_stack;
-pub mod pixel_env;
-pub mod pixel_header;
-pub mod pixel_state;
+pub mod mods;
+pub mod pixels;
 pub mod types;
