+++ src/model/document.rs
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::model::mod_stack::ModStack;
-use crate::model::pixel_state::PixelState;
+use crate::model::mods::ModStack;
+use crate::model::pixels::PixelState;
 use crate::model::types::*;
 
 /// Per-document state.
