--- src/model/mod_record.rs
@@ -1,33 +1,0 @@
-// Copyright 2021 Andy King
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use std::sync::Arc;
-
-use crate::model::types::PixelBytes;
-
-/// Modification record. This holds undo state.
-#[derive(Clone, druid::Data)]
-pub struct ModRecord {
-    pub area: druid::Rect,
-    pub bytes: PixelBytes,
-}
-
-impl ModRecord {
-    pub fn new(area: druid::Rect, bytes: Vec<u8>) -> Self {
-        Self {
-            area,
-            bytes: Arc::new(bytes),
-        }
-    }
-}
