--- src/model/pixel_env.rs
@@ -1,32 +1,0 @@
-// Copyright 2021 Andy King
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/// Capture environment.
-pub struct PixelEnv {
-    pub color: druid::Color,
-    pub pos: druid::Point,
-    pub bounds: druid::Rect,
-    pub param: f64,
-}
-
-impl PixelEnv {
-    pub fn new(color: druid::Color, pos: druid::Point, bounds: druid::Rect, param: f64) -> Self {
-        Self {
-            color,
-            pos,
-            bounds,
-            param,
-        }
-    }
-}
